@@ -1,7 +1,7 @@
 package com.elyeproj.networkexperiment
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
 import com.elyeproj.wikipediafetcher.WikipediaFetcher
 import io.reactivex.disposables.Disposables
